Refactor HemeLB GPU code to use float precision for distribution variables
- Changed the type of distribn_t from double to float in units.h to optimize memory usage and performance on GPU.
- Updated multiple kernel launch calls and memory copy operations in lb.hpp to accommodate the new float type for distribn_t.
- Introduced a new Slurm job script (run.slurm) for running HemeLB-GPU with AMD ROCm, including configuration for various test cases and environment setup.

+++ src/units.h
@@ -23,7 +23,7 @@ namespace hemelb
   // can have type site_t, processors proc_t.
   // Any variable whose precision should roughly match that of the lattice sites' velocity
   // distributions can have type distribn_t.
-  typedef double distribn_t;
+  typedef float distribn_t;
   typedef int proc_t;
   typedef int64_t site_t;
   typedef uint64_t sitedata_t;
